move project component into its own directory
